third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/suppress-mutation-events-in-shadow-characterdata-expected.txt

CONSOLE ERROR: line 11: Uncaught TypeError: host.createShadowRoot is not a function
Ensures that DOMCharacterDataModified isn't fired inside shadows.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
